calculate block starts

+++ corundum_jit/src/translation_state.rs
@@ -3,7 +3,7 @@ use std::collections::HashMap;
 
 pub struct TranslationState {
     pub stack: Vec<Value>,
-    pub blocks: HashMap<usize,Ebb>,
+    pub blocks: HashMap<i32,Ebb>,
     pub between_blocks: bool,
 }
 
@@ -24,11 +24,11 @@ impl TranslationState {
         self.stack.pop().expect("Stack underflow")
     }
 
-    pub fn add_block(&mut self, label: usize, block: Ebb) {
+    pub fn add_block(&mut self, label: i32, block: Ebb) {
         self.blocks.insert(label, block);
     }
 
-    pub fn get_block(&mut self, label: usize) -> Ebb {
+    pub fn get_block(&mut self, label: i32) -> Ebb {
         match self.blocks.get(&label) {
             Some(block) => *block,
             None => panic!("Couldn't get block {}", label)
